finish v1

+++ bin/stop_clusters.sh
@@ -2,8 +2,6 @@
  
  echo "Clearing Kafka Logs..."
   rm -rf /tmp/kafka-logs/
-#   rm -rf /tmp/zookeeper/
-#   rm -rf /tmp/zookeeper-2/
   echo "Finished Clearing Kafka Logs!"
   # $DEPLOY_ROOT_DIR/kafka/bin/kafka-server-stop.sh
   value=`jps | grep Quorum`
@@ -13,8 +11,32 @@
   
   for i in "${a[@]}"; do
     if  [[ $i =~ $re ]] ; then
-      echo "Stopping Zookeeper with PID: ${i}"
-      kill -9 $i
+     if ps -p $i > /dev/null
+        then
+          echo "Stopping Zookeeper with PID: ${i}"
+          kill -9 $i
+        fi
     fi
   done
-  exit
+
+value=$(ps -ef | grep 'zookeeper' | awk '{print $2}')
+ for i in "${value[@]}"; do
+    if  [[ $i =~ $re ]] ; then
+      if ps -p $i > /dev/null
+        then
+          echo "Stopping Zookeeper with PID: ${i}"
+          kill -9 $i
+        fi
+    fi
+  done
+echo "Stopping kafka"
+value2=$(ps -ef | grep 'kafka' | awk '{print $2}')
+ for i in "${value2[@]}"; do
+    if  [[ $i =~ $re ]] ; then
+     if ps -p $i > /dev/null
+        then
+          echo "Stopping Kafka with PID: ${i}"
+          kill -9 $i
+        fi
+    fi
+done
